added utils for the json load, save, file_path generation functionality
removed dulpicate codes for json functionalities
