+++ data/filtered_comments_entries.json
@@ -1,138 +1,132 @@
 [
     {
-        "order": 19,
+        "order": 17,
         "title": "AlphaFold 3 predicts the structure and interactions of life's molecules",
-        "total_comments": 440,
-        "points": 1027
-    },
-    {
-        "order": 23,
-        "title": "Microsoft PlayReady – Complete Client Identity Compromise",
-        "total_comments": 111,
-        "points": 154
-    },
-    {
-        "order": 21,
-        "title": "UTC, Tai, and Unix Time (2001)",
-        "total_comments": 96,
-        "points": 92
-    },
-    {
-        "order": 29,
-        "title": "Consistency LLM: converting LLMs to parallel decoders accelerates inference 3.5x",
-        "total_comments": 81,
-        "points": 417
-    },
-    {
-        "order": 6,
-        "title": "Show HN: Exploring HN by mapping and analyzing 40M posts and comments for fun",
-        "total_comments": 68,
-        "points": 182
-    },
-    {
-        "order": 24,
-        "title": "Deaf girl is cured in world first gene therapy trial",
-        "total_comments": 67,
-        "points": 170
-    },
-    {
-        "order": 14,
-        "title": "Tesla Quietly Removes All U.S. Job Postings",
-        "total_comments": 43,
-        "points": 55
-    },
-    {
-        "order": 30,
-        "title": "Temporal Python – A durable, distributed asyncio event loop (2023)",
-        "total_comments": 37,
-        "points": 155
-    },
-    {
-        "order": 12,
-        "title": "Show HN: An SQS Alternative on Postgres",
-        "total_comments": 33,
-        "points": 104
-    },
-    {
-        "order": 1,
-        "title": "Launch HN: Muddy (YC S19) – Multiplayer browser for getting work done",
-        "total_comments": 29,
-        "points": 82
-    },
-    {
-        "order": 10,
-        "title": "Leaked FBI Email Reportedly Shows Desperation to Justify Warrantless Wiretaps",
-        "total_comments": 28,
-        "points": 88
-    },
-    {
-        "order": 2,
-        "title": "Familial Transmission of Personality Is Higher Than Shown in Typical Studies",
-        "total_comments": 21,
-        "points": 61
+        "total_comments": 442,
+        "points": 1035
     },
     {
         "order": 27,
-        "title": "Andrew Huberman's Claims About Cannabis Annoy Drug Experts",
-        "total_comments": 21,
-        "points": 40
+        "title": "Microsoft PlayReady – Complete Client Identity Compromise",
+        "total_comments": 123,
+        "points": 159
     },
     {
-        "order": 9,
-        "title": "Show HN: Ellipsis – Automated PR reviews and bug fixes",
-        "total_comments": 15,
-        "points": 14
+        "order": 19,
+        "title": "UTC, Tai, and Unix Time (2001)",
+        "total_comments": 98,
+        "points": 96
     },
     {
-        "order": 16,
-        "title": "Show HN: Browser-based knitting (pattern) software",
-        "total_comments": 13,
-        "points": 75
-    },
-    {
-        "order": 22,
-        "title": "Show HN: An open source alternative to some of Slack AI's premium features",
-        "total_comments": 11,
-        "points": 38
-    },
-    {
-        "order": 18,
-        "title": "VideoPrism: A foundational visual encoder for video understanding",
-        "total_comments": 10,
-        "points": 49
-    },
-    {
-        "order": 7,
-        "title": "Leaked Deck Reveals How OpenAI Is Pitching Publisher Partnerships",
-        "total_comments": 8,
-        "points": 8
-    },
-    {
-        "order": 17,
-        "title": "A Brief History of the First Planetarium",
-        "total_comments": 7,
-        "points": 15
-    },
-    {
-        "order": 26,
-        "title": "The Woman Who Showed President Biden ChatGPT",
-        "total_comments": 4,
-        "points": 10
+        "order": 29,
+        "title": "Opening Windows in Linux with sockets, bare hands and 200 lines of C",
+        "total_comments": 94,
+        "points": 206
     },
     {
         "order": 20,
-        "title": "Buckets of Parquet Files Are Awful",
-        "total_comments": 3,
-        "points": 5
+        "title": "Deaf girl is cured in world first gene therapy trial",
+        "total_comments": 88,
+        "points": 202
     },
     {
-        "order": 15,
-        "title": "Show HN: Timmerse.com (YC W22) Easily create immersive real-time interactions",
-        "total_comments": 2,
-        "points": 13
+        "order": 11,
+        "title": "Familial Transmission of Personality Is Higher Than Shown in Typical Studies",
+        "total_comments": 82,
+        "points": 84
+    },
+    {
+        "order": 24,
+        "title": "Consistency LLM: converting LLMs to parallel decoders accelerates inference 3.5x",
+        "total_comments": 81,
+        "points": 422
     },
     {
         "order": 8,
+        "title": "Show HN: Exploring HN by mapping and analyzing 40M posts and comments for fun",
+        "total_comments": 75,
+        "points": 197
+    },
+    {
+        "order": 1,
+        "title": "Leaked Deck Reveals How OpenAI Is Pitching Publisher Partnerships",
+        "total_comments": 40,
+        "points": 73
+    },
+    {
+        "order": 26,
+        "title": "Temporal Python – A durable, distributed asyncio event loop (2023)",
+        "total_comments": 38,
+        "points": 162
+    },
+    {
+        "order": 2,
+        "title": "Launch HN: Muddy (YC S19) – Multiplayer browser for getting work done",
+        "total_comments": 36,
+        "points": 99
+    },
+    {
+        "order": 10,
+        "title": "Show HN: An SQS Alternative on Postgres",
+        "total_comments": 34,
+        "points": 118
+    },
+    {
+        "order": 21,
+        "title": "The Woman Who Showed President Biden ChatGPT",
+        "total_comments": 19,
+        "points": 21
+    },
+    {
+        "order": 7,
+        "title": "Show HN: Ellipsis – Automated PR reviews and bug fixes",
+        "total_comments": 17,
+        "points": 23
+    },
+    {
+        "order": 15,
+        "title": "Show HN: Browser-based knitting (pattern) software",
+        "total_comments": 16,
+        "points": 83
+    },
+    {
+        "order": 23,
+        "title": "Show HN: An open source alternative to some of Slack AI's premium features",
+        "total_comments": 14,
+        "points": 43
+    },
+    {
+        "order": 13,
+        "title": "VideoPrism: A foundational visual encoder for video understanding",
+        "total_comments": 10,
+        "points": 57
+    },
+    {
+        "order": 3,
+        "title": "Shane Legg: Machine Super Intelligence (2008) [pdf]",
+        "total_comments": 9,
+        "points": 24
+    },
+    {
+        "order": 5,
+        "title": "It's always TCP_NODELAY. Every damn time",
+        "total_comments": 7,
+        "points": 19
+    },
+    {
+        "order": 18,
+        "title": "A Brief History of the First Planetarium",
+        "total_comments": 7,
+        "points": 17
+    },
+    {
+        "order": 12,
+        "title": "Show HN: Timmerse.com (YC W22) Easily create immersive real-time interactions",
+        "total_comments": 3,
+        "points": 18
+    },
+    {
+        "order": 9,
         "title": "Freshpaint (YC S19) is hiring software engineers to protect patient privacy",
         "total_comments": 0,
         "points": 0
