+++ data/filtered_points_entries.json
@@ -1,44 +1,50 @@
 [
     {
-        "order": 5,
+        "order": 6,
         "title": "Algebraic Data Types for C99",
-        "total_comments": 113,
-        "points": 226
+        "total_comments": 126,
+        "points": 238
     },
     {
-        "order": 13,
+        "order": 16,
         "title": "No \"Zero-Shot\" Without Exponential Data",
-        "total_comments": 49,
-        "points": 105
-    },
-    {
-        "order": 25,
-        "title": "Xilinx HBM2 Internals (2023)",
-        "total_comments": 9,
-        "points": 52
-    },
-    {
-        "order": 3,
-        "title": "ESP32 Drum Synth Machine",
-        "total_comments": 5,
-        "points": 42
+        "total_comments": 67,
+        "points": 122
     },
     {
         "order": 28,
+        "title": "Symbolica Computer Algebra System",
+        "total_comments": 86,
+        "points": 109
+    },
+    {
+        "order": 4,
+        "title": "ESP32 Drum Synth Machine",
+        "total_comments": 6,
+        "points": 61
+    },
+    {
+        "order": 22,
+        "title": "Xilinx HBM2 Internals (2023)",
+        "total_comments": 10,
+        "points": 57
+    },
+    {
+        "order": 25,
         "title": "Hymn for Walpurgisnacht",
         "total_comments": 12,
         "points": 38
     },
     {
-        "order": 11,
-        "title": "Blade Runner",
-        "total_comments": 3,
-        "points": 12
+        "order": 14,
+        "title": "Logarithmic Scales",
+        "total_comments": 0,
+        "points": 10
     },
     {
-        "order": 4,
-        "title": "Machine Super Intelligence [pdf]",
+        "order": 30,
+        "title": "Sean Baxter: Safe C++ [video]",
         "total_comments": 0,
-        "points": 11
+        "points": 6
     }
 ]
