+++ filter_utilities.py
@@ -1,16 +1,5 @@
-import json
 import os
-
-
-def load_data(file_path):
-    with open(file_path, "r", encoding="utf-8") as f:
-        data = json.load(f)
-    return data
-
-
-def save_data(data, file_path):
-    with open(file_path, "w", encoding="utf-8") as f:
-        json.dump(data, f, ensure_ascii=False, indent=4)
+from utils.json_utils import load_data, save_data, generate_file_path
 
 
 def filter_and_sort_entries_by_comments(entries):
@@ -36,21 +25,15 @@ def filter_and_sort_entries_by_points(entries):
 
 
 def main():
-    original_file = os.path.join("data", "entries.json")
-    comments_output_file = os.path.join("data", "filtered_comments_entries.json")
-    points_output_file = os.path.join("data", "filtered_points_entries.json")
-
+    original_file = generate_file_path("entries.json")
     original_data = load_data(original_file)
+
     filtered_comments_entries = filter_and_sort_entries_by_comments(original_data)
-    print(
-        f"### There are {len(filtered_comments_entries)} entries with more than 5 words ###"
-    )
+    comments_output_file = generate_file_path("filtered_comments_entries.json")
     save_data(filtered_comments_entries, comments_output_file)
 
     filtered_points_entries = filter_and_sort_entries_by_points(original_data)
-    print(
-        f"### There are {len(filtered_points_entries)} entries with less or equal 5 words ###"
-    )
+    points_output_file = generate_file_path("filtered_points_entries.json")
     save_data(filtered_points_entries, points_output_file)
