+++ pipelines.py
@@ -1,5 +1,4 @@
-import json
-import os
+from utils.json_utils import save_data, generate_file_path
 
 class JsonExportPipeline:
     def __init__(self):
@@ -10,6 +9,5 @@ class JsonExportPipeline:
         return item
 
     def close_spider(self, spider):
-        file_path = os.path.join('data', 'entries.json')
-        with open(file_path, "w", encoding="utf-8") as f:
-            json.dump(self.items, f, ensure_ascii=False, indent=4)
+        file_path = generate_file_path(filename='entries.json')
+        save_data(self.items, file_path)
