+++ utils/json_utils.py
@@ -1,0 +1,14 @@
+import json
+import os
+
+def load_data(file_path):
+    with open(file_path, "r", encoding="utf-8") as f:
+        data = json.load(f)
+    return data
+
+def save_data(data, file_path):
+    with open(file_path, "w", encoding="utf-8") as f:
+        json.dump(data, f, ensure_ascii=False, indent=4)
+
+def generate_file_path(filename, folder="data"):
+    return os.path.join(folder, filename)
